+++ fe/src/components/StudentCard.vue
