+++ fe/src/features/authentication/SignupForm.vue
@@ -111,7 +111,7 @@
 import {useQuasar} from 'quasar'
 import {defineComponent, ref} from 'vue'
 import { useMeta } from 'quasar'
-import GitHubSignInButton from '../tools/GitHubSignInButton.vue'
+import GitHubSignInButton from './components/GitHubSignInButton.vue'
 const metaData = {
   title: 'Sign Up',
 }
