+++ fe/src/features/authentication/components/LightDarkButton.vue
