+++ fe/src/features/authentication/components/OsocLogo.vue
@@ -1,6 +1,6 @@
 <template>
-  <img v-if="$q.dark.isActive" src="../../assets/logo_dark.svg"/>
-  <img v-else src="../../assets/logo.svg"/>
+  <img v-if="$q.dark.isActive" src="../../../assets/logo_dark.svg"/>
+  <img v-else src="../../../assets/logo.svg"/>
 </template>
 
 <style scoped>
