+++ fe/src/features/students/components/DetailsCard.vue
