+++ fe/src/features/students/components/PracticalCard.vue
