+++ fe/src/features/students/components/StudentCard.vue
