+++ fe/src/features/students/components/SuggestionsCard.vue
